requirements.txt
telethon
python-decouple
hachoir
ffmpeg-python
motor
dnspython